Network.HTTP: Switch to functional namespace instead of import
